@@ -10,22 +10,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
+    "@angular/animations": "^14.2.12",
     "@angular/cdk": "^13.3.5",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/localize": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
-    "@ng-bootstrap/ng-bootstrap": "^12.0.2",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/localize": "^14.2.12",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
+    "@ng-bootstrap/ng-bootstrap": "^13.1.1",
     "@popperjs/core": "^2.10.2",
     "@types/plotly.js-dist-min": "^2.3.0",
     "angular-plotly.js": "^4.0.4",
     "biomsa": "^0.2.4",
     "bootstrap": "^5.1.3",
+    "curtain-web-api": "^1.0.8",
     "cytoscape": "^3.21.1",
     "data-forge": "^1.9.5",
     "jstat": "^1.9.5",
@@ -37,22 +38,25 @@
     "quill": "^1.3.7",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
+    "uniprotparserjs": "^1.0.18",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.3",
-    "@angular/cli": "~13.3.3",
-    "@angular/compiler-cli": "~13.3.0",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
     "@types/cytoscape": "^3.19.4",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
+    "@types/pouchdb": "^6.4.0",
+    "@types/quill": "^1.3.10",
     "jasmine-core": "~4.0.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.6.2",
-    "@types/quill": "^1.3.10"
+    "pouchdb": "^8.0.1",
+    "typescript": "~4.6.2"
   }
 }
