@@ -8,8 +8,8 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { ToastContainerComponent } from './components/toast-container/toast-container.component';
 import { FileFormComponent } from './components/file-form/file-form.component';
 import {FileInputWidgetComponent} from "./components/file-input-widget/file-input-widget.component";
-import {FormsModule} from "@angular/forms";
-import {HttpClient, HttpClientModule} from "@angular/common/http";
+import {FormsModule, ReactiveFormsModule} from "@angular/forms";
+import {HTTP_INTERCEPTORS, HttpClient, HttpClientModule} from "@angular/common/http";
 import {VolcanoColorsComponent} from "./components/volcano-colors/volcano-colors.component";
 import {VolcanoPlotComponent} from "./components/volcano-plot/volcano-plot.component";
 import {BatchSearchComponent} from "./components/batch-search/batch-search.component";
@@ -38,6 +38,15 @@ import {SampleAnnotationComponent} from "./components/sample-annotation/sample-a
 import {PrideComponent} from "./components/pride/pride.component";
 import { SampleOrderAndHideComponent } from './components/sample-order-and-hide/sample-order-and-hide.component';
 import { VolcanoPlotTextAnnotationComponent } from './components/volcano-plot-text-annotation/volcano-plot-text-annotation.component';
+import { LoginModalComponent } from './accounts/login-modal/login-modal.component';
+import { SessionSettingsComponent } from './components/session-settings/session-settings.component';
+import {AccountsModule} from "./accounts/accounts.module";
+import { KinaseLibraryModalComponent } from './components/kinase-library-modal/kinase-library-modal.component';
+import { WebLogoComponent } from './components/web-logo/web-logo.component';
+import { DefaultColorPaletteComponent } from './components/default-color-palette/default-color-palette.component';
+import { VariantSelectionComponent } from './components/variant-selection/variant-selection.component';
+import { SessionExpiredModalComponent } from './components/session-expired-modal/session-expired-modal.component';
+import { DataSelectionManagementComponent } from './components/data-selection-management/data-selection-management.component';
 PlotlyModule.plotlyjs = PlotlyJS;
 @NgModule({
   declarations: [
@@ -67,7 +76,14 @@ PlotlyModule.plotlyjs = PlotlyJS;
     SampleAnnotationComponent,
     PrideComponent,
     SampleOrderAndHideComponent,
-    VolcanoPlotTextAnnotationComponent
+    VolcanoPlotTextAnnotationComponent,
+    SessionSettingsComponent,
+    KinaseLibraryModalComponent,
+    WebLogoComponent,
+    DefaultColorPaletteComponent,
+    VariantSelectionComponent,
+    SessionExpiredModalComponent,
+    DataSelectionManagementComponent
   ],
   imports: [
     BrowserModule,
@@ -80,6 +96,8 @@ PlotlyModule.plotlyjs = PlotlyJS;
     //ContextMenuModule,
     NgxPrintModule,
     QuillModule.forRoot(),
+    ReactiveFormsModule,
+    AccountsModule
   ],
   providers: [HttpClient,
     //ContextMenuService
