@@ -11,6 +11,7 @@ import {WebService} from "../../web.service";
 import {
   VolcanoPlotTextAnnotationComponent
 } from "../volcano-plot-text-annotation/volcano-plot-text-annotation.component";
+import {WebLogoComponent} from "../web-logo/web-logo.component";
 
 @Component({
   selector: 'app-volcano-plot',
@@ -20,11 +21,11 @@ import {
 export class VolcanoPlotComponent implements OnInit {
   @Output() selected: EventEmitter<selectionData> = new EventEmitter<selectionData>()
   _data: any;
-  nameToID: any = {}
+  //nameToID: any = {}
   graphData: any[] = []
   graphLayout: any = {
-    height: 700, width: 700, xaxis: {title: "Log2FC"},
-    yaxis: {title: "-log10(p-value)"},
+    height: 700, width: 700, xaxis: {title: "<b>Log2FC</b>"},
+    yaxis: {title: "<b>-log10(p-value)</b>"},
     annotations: [],
     showlegend: true, legend: {
       orientation: 'h'
@@ -36,7 +37,16 @@ export class VolcanoPlotComponent implements OnInit {
       },
     }
   }
-
+  config: any = {
+    //modeBarButtonsToRemove: ["toImage"]
+    toImageButtonOptions: {
+      format: 'svg',
+      filename: this.graphLayout.title.text,
+      height: this.graphLayout.height,
+      width: this.graphLayout.width,
+      scale: 1
+    }
+  }
   layoutMaxMin: any = {
     xMin: 0, xMax: 0, yMin: 0, yMax: 0
   }
@@ -55,8 +65,12 @@ export class VolcanoPlotComponent implements OnInit {
   }
 
   breakColor: boolean = false
-
+  markerSize: number = 10
   drawVolcano() {
+    if (!this.settings.settings.visible) {
+      this.settings.settings.visible = {}
+    }
+    this.settings.settings.scatterPlotMarkerSize = this.markerSize
     this.graphLayout.title.text = this.settings.settings.volcanoPlotTitle
     let currentColors: string[] = []
     if (this.settings.settings.colorMap) {
@@ -77,13 +91,13 @@ export class VolcanoPlotComponent implements OnInit {
       if (!this.settings.settings.colorMap[s]) {
         while (true) {
           if (this.breakColor) {
-            this.settings.settings.colorMap[s] = this.dataService.defaultColorList[currentPosition]
+            this.settings.settings.colorMap[s] = this.settings.settings.defaultColorList[currentPosition]
             break
           }
-          if (currentColors.indexOf(this.dataService.defaultColorList[currentPosition]) !== -1) {
+          if (currentColors.indexOf(this.settings.settings.defaultColorList[currentPosition]) !== -1) {
             currentPosition ++
-          } else if (currentPosition !== this.dataService.defaultColorList.length) {
-            this.settings.settings.colorMap[s] = this.dataService.defaultColorList[currentPosition]
+          } else if (currentPosition !== this.settings.settings.defaultColorList.length) {
+            this.settings.settings.colorMap[s] = this.settings.settings.defaultColorList[currentPosition]
             break
           } else {
             this.breakColor = true
@@ -92,7 +106,7 @@ export class VolcanoPlotComponent implements OnInit {
         }
 
         currentPosition ++
-        if (currentPosition === this.dataService.defaultColorList.length) {
+        if (currentPosition === this.settings.settings.defaultColorList.length) {
           currentPosition = 0
         }
       }
@@ -101,15 +115,19 @@ export class VolcanoPlotComponent implements OnInit {
         x: [],
         y: [],
         text: [],
-        type: "scattergl",
+        primaryIDs: [],
+        //type: "scattergl",
+        type: "scatter",
         mode: "markers",
         name: s,
         marker: {
-          color: this.settings.settings.colorMap[s]
+          color: this.settings.settings.colorMap[s],
+          size: this.settings.settings.scatterPlotMarkerSize
         }
       }
 
     }
+    console.log(this.settings.settings.colorMap)
     this.layoutMaxMin = {
       xMin: 0, xMax: 0, yMin: 0, yMax: 0
     }
@@ -136,7 +154,9 @@ export class VolcanoPlotComponent implements OnInit {
       x:[],
       y:[],
       text: [],
-      type: "scattergl",
+      primaryIDs: [],
+      //type: "scattergl",
+      type: "scatter",
       mode: "markers",
       name: "Background"
     }
@@ -144,6 +164,7 @@ export class VolcanoPlotComponent implements OnInit {
       temp["Background"]["marker"] = {
         color: "#a4a2a2",
         opacity: 0.3,
+        size: this.settings.settings.scatterPlotMarkerSize
       }
     }
     for (const r of this._data) {
@@ -154,7 +175,7 @@ export class VolcanoPlotComponent implements OnInit {
       const accID = r[this.dataService.differentialForm.accession]
       let text = primaryID
       if (this.dataService.fetchUniProt) {
-        const r = this.uniprot.getUniprotFromAcc(accID)
+        const r: any = this.uniprot.getUniprotFromAcc(accID)
         if (r) {
           geneNames = r["Gene Names"]
         }
@@ -166,24 +187,26 @@ export class VolcanoPlotComponent implements OnInit {
       if (geneNames !== "") {
         text = geneNames + "(" + primaryID + ")"
       }
-      this.nameToID[text] = primaryID
+      //this.nameToID[text] = primaryID
       if (this.dataService.selectedMap[primaryID]) {
         for (const o in this.dataService.selectedMap[primaryID]) {
           temp[o].x.push(x)
           temp[o].y.push(y)
           temp[o].text.push(text)
+          temp[o].primaryIDs.push(primaryID)
         }
       } else if (this.settings.settings.backGroundColorGrey) {
         temp["Background"].x.push(x)
         temp["Background"].y.push(y)
         temp["Background"].text.push(text)
+        temp["Background"].primaryIDs.push(primaryID)
       } else {
         const group = this.dataService.significantGroup(x, y)
         if (!temp[group]) {
           if (!this.settings.settings.colorMap[group]) {
-            this.settings.settings.colorMap[group] = this.dataService.defaultColorList[currentPosition]
+            this.settings.settings.colorMap[group] = this.settings.settings.defaultColorList[currentPosition]
             currentPosition ++
-            if (currentPosition === this.dataService.defaultColorList.length) {
+            if (currentPosition === this.settings.settings.defaultColorList.length) {
               currentPosition = 0
             }
           }
@@ -192,10 +215,13 @@ export class VolcanoPlotComponent implements OnInit {
             x: [],
             y: [],
             text: [],
-            type: "scattergl",
+            primaryIDs: [],
+            //type: "scattergl",
+            type: "scatter",
             mode: "markers",
             marker: {
               color: this.settings.settings.colorMap[group],
+              size: this.settings.settings.scatterPlotMarkerSize
             },
             name: group
           }
@@ -203,13 +229,20 @@ export class VolcanoPlotComponent implements OnInit {
         temp[group].x.push(x)
         temp[group].y.push(y)
         temp[group].text.push(text)
+        temp[group].primaryIDs.push(primaryID)
       }
     }
-    console.log(temp)
     const graphData: any[] = []
     for (const t in temp) {
       if (temp[t].x.length > 0) {
-        graphData.push(temp[t])
+        if (temp[t].x.length > 0) {
+          if (this.settings.settings.visible[t]) {
+            temp[t].visible = this.settings.settings.visible[t]
+          } else {
+            temp[t].visible = true
+          }
+          graphData.push(temp[t])
+        }
       }
     }
     if (fdrCurve.count() > 0) {
@@ -323,29 +356,55 @@ export class VolcanoPlotComponent implements OnInit {
     }
     this.graphData = graphData.reverse()
     this.graphLayout.annotations = []
+    console.log(this.settings.settings.textAnnotation)
     for (const i in this.settings.settings.textAnnotation) {
-      this.annotated[this.settings.settings.textAnnotation[i].title] = this.settings.settings.textAnnotation[i].data
-      this.graphLayout.annotations.push(this.settings.settings.textAnnotation[i].data)
+      if (this.settings.settings.textAnnotation[i].showannotation === true) {
+        this.annotated[this.settings.settings.textAnnotation[i].title] = this.settings.settings.textAnnotation[i].data
+        this.graphLayout.annotations.push(this.settings.settings.textAnnotation[i].data)
+      }
+
     }
+    console.log(this.graphLayout.annotations)
   }
 
   constructor(private web: WebService, private dataService: DataService, private uniprot: UniprotService, public settings: SettingsService, private modal: NgbModal) {
-    this.annotated = this.dataService.annotatedData
+    this.annotated = {}
+    for (const i in this.settings.settings.textAnnotation) {
+      if (this.settings.settings.textAnnotation[i].showannotation === undefined || this.settings.settings.textAnnotation[i].showannotation === null) {
+        this.settings.settings.textAnnotation[i].showannotation = true
+      }
+      this.annotated[i] = this.settings.settings.textAnnotation[i]
+    }
     this.dataService.selectionUpdateTrigger.asObservable().subscribe(data => {
       if (data) {
         this.drawVolcano()
       }
     })
     this.dataService.annotationService.asObservable().subscribe(data => {
-      console.log(data)
       if (data) {
         if (data.remove) {
-          this.removeAnnotatedDataPoints([data.id])
+          if (typeof data.id === "string") {
+            this.removeAnnotatedDataPoints([data.id]).then(() => {
+              this.dataService.annotatedData = this.annotated
+            })
+          } else {
+            this.removeAnnotatedDataPoints(data.id).then(() => {
+              this.dataService.annotatedData = this.annotated
+            })
+          }
+
         } else {
-          this.annotateDataPoints([data.id])
+          if (typeof data.id === "string") {
+            this.annotateDataPoints([data.id]).then(() => {
+              this.dataService.annotatedData = this.annotated
+            })
+          } else {
+            this.annotateDataPoints(data.id).then(() => {
+              this.dataService.annotatedData = this.annotated
+            })
+          }
+
         }
-        console.log(this.annotated)
-        this.dataService.annotatedData = this.annotated
       }
     })
   }
@@ -357,9 +416,7 @@ export class VolcanoPlotComponent implements OnInit {
     if ("points" in e) {
       const selected: string[] = []
       for (const p of e["points"]) {
-        if (this.nameToID[p.text] !== "") {
-          selected.push(this.nameToID[p.text])
-        }
+        selected.push(p.data.primaryIDs[p.pointNumber])
       }
       if (selected.length === 1) {
         this.selected.emit(
@@ -388,12 +445,12 @@ export class VolcanoPlotComponent implements OnInit {
     this.modal.open(VolcanoColorsComponent)
   }
 
-  annotateDataPoints(data: string[]) {
+  async annotateDataPoints(data: string[]) {
     const annotations: any[] = []
     const annotatedData = this.dataService.currentDF.where(r => data.includes(r[this.dataService.differentialForm.primaryIDs])).bake()
     for (const a of annotatedData) {
       let title = a[this.dataService.differentialForm.primaryIDs]
-      const uni = this.uniprot.getUniprotFromAcc(a[this.dataService.differentialForm.primaryIDs])
+      const uni: any = this.uniprot.getUniprotFromAcc(a[this.dataService.differentialForm.primaryIDs])
       if (uni) {
         if (uni["Gene Names"] !== "") {
           title = uni["Gene Names"] + "(" + title + ")"
@@ -423,27 +480,26 @@ export class VolcanoPlotComponent implements OnInit {
           this.settings.settings.textAnnotation[title] = {
             primary_id: a[this.dataService.differentialForm.primaryIDs],
             data: ann,
-            title: title
+            title: title,
+            showannotation: true
           }
         }
         annotations.push(ann)
         this.annotated[title] = ann
       }
     }
-    console.log(this.annotated)
+
     if (annotations.length > 0) {
-      console.log(annotations)
       this.graphLayout.annotations = annotations.concat(this.graphLayout.annotations)
       console.log(this.graphLayout.annotations)
     }
-    console.log(this.annotated)
   }
 
-  removeAnnotatedDataPoints(data: string[]) {
+  async removeAnnotatedDataPoints(data: string[]) {
     const annotatedData = this.dataService.currentDF.where(r => data.includes(r[this.dataService.differentialForm.primaryIDs])).bake()
     for (const d of annotatedData) {
       let title = d[this.dataService.differentialForm.primaryIDs]
-      const uni = this.uniprot.getUniprotFromAcc(d[this.dataService.differentialForm.primaryIDs])
+      const uni:any = this.uniprot.getUniprotFromAcc(d[this.dataService.differentialForm.primaryIDs])
       if (uni) {
         if (uni["Gene Names"] !== "") {
           title = uni["Gene Names"] + "(" + title + ")"
@@ -451,6 +507,7 @@ export class VolcanoPlotComponent implements OnInit {
       }
       if (this.annotated[title]) {
         delete this.annotated[title]
+        delete this.settings.settings.textAnnotation[title]
       }
     }
     this.graphLayout.annotations = Object.values(this.annotated)
@@ -461,15 +518,37 @@ export class VolcanoPlotComponent implements OnInit {
   }
 
   openTextEditor() {
-    const ref = this.modal.open(VolcanoPlotTextAnnotationComponent, {size: "xl"})
+    const ref = this.modal.open(VolcanoPlotTextAnnotationComponent, {size: "xl", scrollable: true})
     ref.closed.subscribe(data => {
       this.graphLayout.annotations = []
       this.annotated = {}
-      for (const a of data) {
-        this.settings.settings.textAnnotation[a.title] = a
-        this.annotated[a.title] = a.data
-        this.graphLayout.annotations.push(this.annotated[a.title])
+      for (const f of data) {
+        this.settings.settings.textAnnotation[f.value.annotationID].data.showarrow = f.value.showarrow
+        this.settings.settings.textAnnotation[f.value.annotationID].data.arrowhead = f.value.arrowhead
+        this.settings.settings.textAnnotation[f.value.annotationID].data.arrowsize = f.value.arrowsize
+        this.settings.settings.textAnnotation[f.value.annotationID].data.arrowwidth = f.value.arrowwidth
+        this.settings.settings.textAnnotation[f.value.annotationID].data.ax = f.value.ax
+        this.settings.settings.textAnnotation[f.value.annotationID].data.ay = f.value.ay
+        this.settings.settings.textAnnotation[f.value.annotationID].data.font.size = f.value.fontsize
+        this.settings.settings.textAnnotation[f.value.annotationID].data.font.color = f.value.fontcolor
+        this.settings.settings.textAnnotation[f.value.annotationID].data.text = f.value.text
+        this.settings.settings.textAnnotation[f.value.annotationID].showannotation = f.value.showannotation
+        this.annotated[f.value.annotationID] = this.settings.settings.textAnnotation[f.value.annotationID].data
+        this.graphLayout.annotations.push(this.annotated[f.value.annotationID])
       }
     })
+  }
+
+  legendClickHandler(event: any) {
+    if (event.event.srcElement.__data__[0].trace.visible === "legendonly") {
+      this.settings.settings.visible[event.event.srcElement.__data__[0].trace.name] = true
+    } else {
+      this.settings.settings.visible[event.event.srcElement.__data__[0].trace.name] = "legendonly"
+    }
+  }
+
+  openWebLogo() {
+    const ref = this.modal.open(WebLogoComponent, {size: "lg"})
+    ref.componentInstance.data = this.dataService.currentDF.where(r => r[this.dataService.differentialForm.primaryIDs] in this.dataService.selectedMap).bake()
   }
 }
