@@ -11,11 +11,15 @@ import {UniprotService} from "./uniprot.service";
   providedIn: 'root'
 })
 export class DataService {
+  session: any = {}
+  tempLink: boolean = false
   dataClear: Subject<boolean> = new Subject()
   finishedProcessingData: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false)
   selectionUpdateTrigger: Subject<boolean> = new Subject<boolean>()
   restoreTrigger: Subject<boolean> = new Subject<boolean>()
   annotationService: Subject<any> = new Subject<any>()
+  batchAnnotateAnnoucement: Subject<any> = new Subject<any>()
+  updateVariantCorrection: Subject<boolean> = new Subject<boolean>()
   raw: InputFile = new InputFile()
   rawForm: Raw = new Raw()
   differential: InputFile = new InputFile()
@@ -33,6 +37,97 @@ export class DataService {
     "ANOVA",
     //"TTest"
   ]
+  palette: any = {
+    "pastel": [
+      "#fd7f6f",
+      "#7eb0d5",
+      "#b2e061",
+      "#bd7ebe",
+      "#ffb55a",
+      "#ffee65",
+      "#beb9db",
+      "#fdcce5",
+      "#8bd3c7"
+    ], "retro": [
+      "#ea5545",
+      "#f46a9b",
+      "#ef9b20",
+      "#edbf33",
+      "#ede15b",
+      "#bdcf32",
+      "#87bc45",
+      "#27aeef",
+      "#b33dc6"
+    ],
+    "solid": [
+      '#1f77b4',
+      '#ff7f0e',
+      '#2ca02c',
+      '#d62728',
+      '#9467bd',
+      '#8c564b',
+      '#e377c2',
+      '#7f7f7f',
+      '#bcbd22',
+      '#17becf'
+    ],
+    "gradient_red_to_green": [
+      "#ff0000",
+      "#ff3300",
+      "#ff6600",
+      "#ff9900",
+      "#ffcc00",
+      "#ffff00",
+      "#ccff00",
+      "#99ff00",
+      "#66ff00",
+      "#33ff00",
+      "#00ff00"
+    ],
+    "Tol_bright": [
+      '#EE6677',
+      '#228833',
+      '#4477AA',
+      '#CCBB44',
+      '#66CCEE',
+      '#AA3377',
+      '#BBBBBB'
+    ],
+    "Tol_muted": [
+      '#88CCEE',
+      '#44AA99',
+      '#117733',
+      '#332288',
+      '#DDCC77',
+      '#999933',
+      '#CC6677',
+      '#882255',
+      '#AA4499',
+      '#DDDDDD'
+    ],
+    "Tol_light": [
+      '#BBCC33',
+      '#AAAA00',
+      '#77AADD',
+      '#EE8866',
+      '#EEDD88',
+      '#FFAABB',
+      '#99DDFF',
+      '#44BB99',
+      '#DDDDDD'
+    ],
+    "Okabe_Ito": [
+      "#E69F00",
+      "#56B4E9",
+      "#009E73",
+      "#F0E442",
+      "#0072B2",
+      "#D55E00",
+      "#CC79A7",
+      "#000000"
+    ]
+  }
+
   currentDF: IDataFrame = new DataFrame()
   accessionToPrimaryIDs: any = {}
   primaryIDsList: string[] = []
@@ -50,9 +145,10 @@ export class DataService {
   selectedMap: any = {}
   selectOperationNames: string[] = []
   searchType: string = "Gene Names"
-  page: number = 0
+  page: number = 1
+  pageSize: number = 5
   dbIDMap: any = {}
-  defaultColorList = [
+  /*defaultColorList = [
     '#1f77b4',
     '#ff7f0e',
     '#2ca02c',
@@ -62,7 +158,18 @@ export class DataService {
     '#e377c2',
     '#7f7f7f',
     '#bcbd22',
-    '#17becf']
+    '#17becf']*/
+  defaultColorList = [
+    "#fd7f6f",
+    "#7eb0d5",
+    "#b2e061",
+    "#bd7ebe",
+    "#ffb55a",
+    "#ffee65",
+    "#beb9db",
+    "#fdcce5",
+    "#8bd3c7"
+  ]
   redrawTrigger: Subject<boolean> = new Subject()
   colorMap: any = {}
   constructor(private settings: SettingsService, private uniprot: UniprotService) { }
@@ -71,9 +178,9 @@ export class DataService {
     const ylog = -Math.log10(this.settings.settings.pCutoff)
     const groups: string[] = []
     if (ylog > y) {
-      groups.push("P-value < " + this.settings.settings.pCutoff)
+      groups.push("P-value > " + this.settings.settings.pCutoff)
     } else {
-      groups.push("P-value >= " + this.settings.settings.pCutoff)
+      groups.push("P-value <= " + this.settings.settings.pCutoff)
     }
 
     if (Math.abs(x) > this.settings.settings.log2FCCutoff) {
