@@ -1,4 +1,6 @@
 export const environment = {
   production: true,
-  apiURL: "http://curtain.omics.quest/"
+  apiURL: "https://celsus.muttsu.xyz/",
+  google: "101803197167-s9fb349oeqhjmvs5afn9idopgipnlvtk.apps.googleusercontent.com",
+  orcid: "APP-5RXLC3W1MS2MOW0F"
 };
