@@ -4,7 +4,10 @@
 
 export const environment = {
   production: false,
-  apiURL: "http://curtain.omics.quest/"
+  apiURL: "https://celsus.muttsu.xyz/",
+  //apiURL: "http://127.0.0.1:8000/",
+  google: "101803197167-s9fb349oeqhjmvs5afn9idopgipnlvtk.apps.googleusercontent.com",
+  orcid: "APP-5RXLC3W1MS2MOW0F"
 };
 
 /*
